Repositories/electricgarden-electricgardenfirmware-3cf1cbd7a6ae/provisioning/requirements.txt
requests

pymongo

pyserial

websocket

python-dotenv